updatePaymentMomo

+++ Backend_ASP.NET/Data/Orders.cs
@@ -6,7 +6,7 @@
     {
         public string? ShipAddress { get; set; }
         public DateTime? ShippperDate { get; set; }
-        public int?  TotalAmount { get; set; }
+        public double?  TotalAmount { get; set; }
         public bool? OrderStatus { get; set; }
         public string? PaymentType { get; set; }
