+++ Backend_ASP.NET/Models/OrderModel.cs
@@ -8,7 +8,7 @@ namespace Backend_ASP.NET.Models
         public Guid Id { get; set; }
         public string? ShipAddress { get; set; }
         public DateTime? ShippperDate { get; set; }
-        public int? TotalAmount { get; set; }
+        public double? TotalAmount { get; set; }
         public bool? OrderStatus { get; set; }
         public string? PaymentType { get; set; }
         public Guid? StoreID { get; set; }
